completed all assignments for exercise, 4 new modules made, tested and working smooth as far as I can tell
