+++ javascripts/components/carttotal.js
@@ -1,0 +1,12 @@
+import { getCart } from './../helpers/cartData.js'
+
+const cartTotal = () => {
+    const myCart = getCart();
+    const total = myCart.reduce((a, cartItem) => {
+      return a + cartItem.price;
+    }, 0);
+  
+    return total;
+  };
+
+  export { cartTotal };
